backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
anthropic==0.28.0
pydantic==2.7.4
python-multipart==0.0.9
PyPDF2==3.0.1
pandas==2.2.2
numpy==1.26.4
scipy==1.15.0
yfinance==0.2.40
httpx==0.27.0
python-dotenv==1.0.1
openai>=1.50.0
